Tweak sizes and colors
Smaller content font size but also narrower containers for more natural
line flow. Also adjusts the vertical spacing between paragraphs to be a
multiple of the line height for a nicer vertical rhythm.

Oh, and I increased the contrast of the headline color so it pops out
just a tiny bit more over the slightly darker content box.